fix styles on input todo and refactor components
