+++ homeworks/federico-benitez/20-routing-redux/todo-task/src/components/item/index.js
@@ -1,0 +1,3 @@
+export { default as EditTodo } from './EditTodo';
+export { default as NewTodo } from './NewTodo';
+export { default as TodoItem } from './TodoItem';
