+++ homeworks/federico-benitez/20-routing-redux/todo-task/src/index.css
@@ -1,5 +1,7 @@
 * {
   box-sizing: border-box;
+  -webkit-box-sizing: border-box; /* Safari/Chrome, other WebKit */
+  -moz-box-sizing: border-box; /* Firefox, other Gecko */
   margin: 0;
   padding: 0;
   font-size: 32px;
@@ -52,6 +54,12 @@
   background: var(--background-input);
   width: 554px;
   height: 65px;
+  color: #00000060;
+  text-align: center;
+  padding: 10px;
+  -webkit-box-sizing: border-box; /* Safari/Chrome, other WebKit */
+  -moz-box-sizing: border-box; /* Firefox, other Gecko */
+  box-sizing: border-box;
 }
 
 .button__add {
